PLANNER-1652 Simplify and move comparator class into the test class
